requirements.txt
appdirs==1.4.4
astroid==2.4.2
autopep8==1.5.5
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
distlib==0.3.1
filelock==3.0.12
flake8==3.8.4
idna==2.10
importlib-metadata==3.6.0
inflection==0.5.1
ipykernel==5.5.0
ipython==7.20.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.18.0
joblib==1.0.1
jupyter-client==6.1.11
jupyter-core==4.7.1
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.3.4
mccabe==0.6.1
more-itertools==8.7.0
numexpr==2.7.2
numpy==1.20.1
pandas==1.2.2
parso==0.8.1
patsy==0.5.1
pickleshare==0.7.5
Pillow==8.1.0
pipenv==2020.6.2
prompt-toolkit==3.0.16
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.8.0
pylint==2.5.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
pywin32==300
pyzmq==22.0.3
Quandl==3.6.0
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
sklearn==0.0
statsmodels==0.12.2
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.3
virtualenv==20.0.25
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
wrapt==1.12.1
zipp==3.4.0
